refactor(jazz-tools): Phase 8 (row_histories) — split mod.rs along type/codec/apply seams
row_histories/mod.rs was ~3.4K LOC mixing data types, descriptor codecs, and
apply-time algorithms. Split into three submodules + a thin re-export root:

  row_histories/
    mod.rs    — module docs, sibling re-exports, and the existing tests block
    types.rs  — BatchId, RowState, HistoryScan, RowVisibilityChange,
                ApplyRowBatchResult, RowHistoryError, QueryRowBatch,
                StoredRowBatch, RowMetadata, VisibleRowEntry (and its impl,
                which still calls into apply-side helpers via pub(super))
    codecs.rs — descriptor builders + flat-row encoders/decoders, plus the
                shared utilities (compute_row_digest, tier_satisfies,
                malformed, flat_user_values, project_user_row_data_from_physical)
    apply.rs  — apply_row_batch, patch_row_batch_state, visible-row
                computation helpers (build_computed_visible_preview,
                latest_common_ancestor, merge_column_with_strategy, …),
                visibility_change_from_applied, branch_frontier,
                latest_visible_version_for_tier

Cross-file helpers are exposed as pub(super) so the surface area beyond the
module stays unchanged; the existing public re-exports in mod.rs keep every
caller (`row_histories::apply_row_batch`, `::compute_row_digest`,
`::StoredRowBatch`, etc.) resolving as before. The `pub(crate)` items
(`flat_row_codecs`, `decode_flat_*_with_codecs`, `FlatRowCodecs`,
`visible_row_preview_from_history_rows`) keep their crate-internal visibility.

Tests stay in `mod.rs` — they're cross-cutting and exercise both codecs
and apply-side behaviour through MemoryStorage.

Independent of every other phase. Spec:
specs/todo/c_later/crate_structure_cleanup.md (Phase 8 — row_histories
half; routes.rs split deferred since it depends on Phase 1 landing).
